requirements.txt
discord.py==2.3.2
asyncio==3.4.3
numpy==1.24.3
torch==2.0.1
openai-whisper==20231117
requests==2.31.0
PyAudio==0.2.14
keyboard==0.13.5
librosa==0.10.1
PyNaCl == 1.5.0